Restruct packages
