+++ exception/src/main/kotlin/validation/Validation.kt
@@ -1,0 +1,7 @@
+package validation
+
+import ErrorCode
+
+interface Validation<E> {
+    fun validate(data: E): ErrorCode?
+}
